lib/posix-user: Separate user and group calls
For a better overview this commit separates functions related to
user/password and group management and moves them into
their own files. The former `user.c` is used as `posix-user` main
that holds the constructor.
