+++ lib/posix-user/Makefile.uk
@@ -5,6 +5,9 @@ LIBPOSIX_USER_COMMON_INCLUDES-y += -I$(LIBPOSIX_USER_BASE)/musl-imported/include
 CINCLUDES-y   += $(LIBPOSIX_USER_COMMON_INCLUDES-y)
 CXXINCLUDES-y += $(LIBPOSIX_USER_COMMON_INCLUDES-y)
 LIBPOSIX_USER_SRCS-$(CONFIG_LIBPOSIX_USER) += $(LIBPOSIX_USER_BASE)/user.c
+LIBPOSIX_USER_SRCS-$(CONFIG_LIBPOSIX_USER) += $(LIBPOSIX_USER_BASE)/group.c
+LIBPOSIX_USER_SRCS-$(CONFIG_LIBPOSIX_USER) += $(LIBPOSIX_USER_BASE)/passwd.c
+
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_USER) += getegid-0
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_USER) += geteuid-0
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_USER) += getgid-0
