+++ lib/posix-user/user.c
@@ -2,8 +2,11 @@
 /*
  * Authors: Felipe Huici <felipe.huici@neclab.eu>
  *          Simon Kuenzer <simon.kuenzer@neclab.eu>
+ *          Marc Rittinghaus <marc.rittinghaus@kit.edu>
  *
  * Copyright (c) 2018, NEC Europe Ltd., NEC Corporation. All rights reserved.
+ * Copyright (c) 2022, Karlsruhe Institute of Technology (KIT).
+ *                     All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -31,384 +34,13 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <unistd.h>
-#include <grp.h>
-#include <pwd.h>
-#include <string.h>
-#include <errno.h>
-#include <sys/types.h>
-#include <uk/essentials.h>
-#include <uk/list.h>
+#include "user.h"
+
 #include <uk/ctors.h>
-#include <uk/print.h>
-#include <uk/user.h>
-#include <uk/syscall.h>
-
-#define UK_DEFAULT_UID    0
-#define UK_DEFAULT_GID    0
-#define UK_DEFAULT_USER   "root"
-#define UK_DEFAULT_GROUP  "root"
-#define UK_DEFAULT_PASS   "x"
-
-static __uk_tls struct passwd_entry {
-	struct passwd *passwd;
-
-	UK_SLIST_ENTRY(struct passwd_entry) entries;
-} *iter;
-
-UK_SLIST_HEAD(uk_entry_list, struct passwd_entry);
-
-static struct uk_entry_list passwds;
-
-static void init_groups(void);
-
-/*
- * TODO make passwd management consistent with group management
- */
 
 static void init_posix_user(void)
 {
-	static struct passwd_entry p1;
-	static struct passwd passwd = {
-		.pw_name = UK_DEFAULT_USER,
-		.pw_passwd = UK_DEFAULT_PASS,
-		.pw_uid = UK_DEFAULT_UID,
-		.pw_gid = UK_DEFAULT_GID,
-		.pw_gecos = UK_DEFAULT_USER,
-		.pw_dir = "/",
-		.pw_shell = NULL,
-	};
-
-	p1.passwd = &passwd;
-
-	UK_SLIST_INIT(&passwds);
-	UK_SLIST_INSERT_HEAD(&passwds, &p1, entries);
-
-	init_groups();
+	pu_init_passwds();
+	pu_init_groups();
 }
 UK_CTOR_FUNC(2, init_posix_user);
-
-UK_SYSCALL_R_DEFINE(uid_t, getuid)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, setuid, uid_t, uid)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(uid_t, geteuid)
-{
-	return 0;
-}
-
-int seteuid(uid_t euid __unused)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, getresuid, uid_t*, ruid, uid_t*, euid, uid_t*, suid)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, setresuid, uid_t, ruid, uid_t, euid, uid_t, suid)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, setreuid, uid_t, ruid, uid_t, euid)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, setfsgid, uid_t, fsgid)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, setfsuid, uid_t, fsuid)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, capset, void*, hdrp, void*, datap)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, capget, void*, hdrp, void*, datap)
-{
-	return 0;
-}
-
-char *getlogin(void)
-{
-	return 0;
-}
-
-void setpwent(void)
-{
-	iter = UK_SLIST_FIRST(&passwds);
-}
-
-void endpwent(void)
-{
-}
-
-struct passwd *getpwnam(const char *name)
-{
-	struct passwd *pwd;
-
-	setpwent();
-	while ((pwd = getpwent()) && strcmp(pwd->pw_name, name))
-		;
-	endpwent();
-
-	return pwd;
-}
-
-struct passwd *getpwuid(uid_t uid)
-{
-	struct passwd *pwd;
-
-	setpwent();
-	while ((pwd = getpwent()) && pwd->pw_uid != uid)
-		;
-	endpwent();
-
-	return pwd;
-}
-
-int getpwnam_r(const char *name __unused, struct passwd *pwd __unused,
-	char *buf __unused, size_t buflen __unused,
-	struct passwd **result __unused)
-{
-	return 0;
-}
-
-int getpwuid_r(uid_t uid __unused, struct passwd *pwd __unused,
-	char *buf __unused, size_t buflen __unused,
-	struct passwd **result __unused)
-{
-	return 0;
-}
-
-struct passwd *getpwent(void)
-{
-	struct passwd *pwd;
-
-	if (iter == NULL)
-		return NULL;
-
-	pwd = iter->passwd;
-
-	iter = UK_SLIST_NEXT(iter, entries);
-
-	return pwd;
-}
-
-UK_SYSCALL_R_DEFINE(gid_t, getgid)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, setgid, gid_t, gid)
-{
-	return 0;
-}
-
-int issetugid(void)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(gid_t, getegid)
-{
-	return 0;
-}
-
-int setegid(gid_t egid __unused)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, getresgid, gid_t*, rgid, gid_t*, egid, gid_t*, sgid)
-{
-	if (!rgid || !egid || !sgid) {
-		return -EFAULT;
-	}
-
-	*rgid = *egid = *sgid = UK_DEFAULT_GID;
-
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, setresgid, gid_t, rgid, gid_t, egid, gid_t, sgid)
-{
-	/* We allow only UK_DEFAULT_GID */
-	if (rgid != UK_DEFAULT_GID || egid != UK_DEFAULT_GID ||
-			sgid != UK_DEFAULT_GID) {
-		return -EINVAL;
-	}
-
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, setregid, gid_t, rgid, gid_t, egid)
-{
-	return 0;
-}
-
-int initgroups(const char *user __unused, gid_t group __unused)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, getgroups, int, size, gid_t*, list)
-{
-	return 0;
-}
-
-UK_SYSCALL_R_DEFINE(int, setgroups, size_t, size, const gid_t*, list)
-{
-	return 0;
-}
-
-/* Group members */
-static char *g_members__[] = { UK_DEFAULT_USER, NULL };
-
-/* Group entry */
-static struct group g__ = {
-	.gr_name = UK_DEFAULT_GROUP,
-	.gr_passwd = UK_DEFAULT_PASS,
-	.gr_gid = UK_DEFAULT_GID,
-	.gr_mem = g_members__,
-};
-
-struct group *getgrnam(const char *name)
-{
-	struct group *res;
-
-	if (name && !strcmp(name, g__.gr_name))
-		res = &g__;
-	else {
-		res = NULL;
-		errno = ENOENT;
-	}
-
-	return res;
-}
-
-int getgrnam_r(const char *name, struct group *grp,
-		char *buf, size_t buflen, struct group **result)
-{
-	size_t needed;
-	int i, members;
-
-	if (!name || strcmp(name, g__.gr_name)) {
-		*result = NULL;
-		return 0;
-	}
-
-	/* check if provided buffer is big enough */
-	needed = strlen(g__.gr_name) + 1
-			+ strlen(g__.gr_passwd) + 1
-			+ sizeof(g__.gr_gid);
-	i = 0;
-	while (g__.gr_mem[i])
-		needed += strlen(g__.gr_mem[i++]) + 1;
-	members = i;
-
-	if (buflen < needed) {
-		*result = NULL;
-		return ERANGE;
-	}
-
-	/* set name */
-	strlcpy(buf, g__.gr_name, strlen(g__.gr_name) + 1);
-	grp->gr_name = buf;
-	buf += strlen(g__.gr_name) + 1;
-
-	/* set passwd */
-	strlcpy(buf, g__.gr_passwd, strlen(g__.gr_passwd) + 1);
-	grp->gr_passwd = buf;
-	buf += strlen(g__.gr_passwd) + 1;
-
-	/* set gid */
-	grp->gr_gid = g__.gr_gid;
-
-	/* set members */
-	i = 0;
-	grp->gr_mem = (char **) buf;
-	buf += (members + 1) * sizeof(char *);
-	while (g__.gr_mem[i]) {
-		strlcpy(buf, g__.gr_mem[i], strlen(g__.gr_mem[i]) + 1);
-		grp->gr_mem[i] = buf;
-		buf += strlen(g__.gr_mem[i]) + 1;
-		i++;
-	}
-	grp->gr_mem[i] = NULL;
-
-	*result = grp;
-
-	return 0;
-}
-
-struct group *getgrgid(gid_t gid)
-{
-	struct group *res;
-
-	if (gid == g__.gr_gid)
-		res = &g__;
-	else {
-		res = NULL;
-		errno = ENOENT;
-	}
-
-	return res;
-}
-
-static __uk_tls struct group_entry {
-	struct group *group;
-
-	UK_SLIST_ENTRY(struct group_entry) entries;
-} *groups_iter;
-
-UK_SLIST_HEAD(uk_group_entry_list, struct group_entry);
-
-static struct uk_group_entry_list groups;
-
-static void init_groups(void)
-{
-	static struct group_entry ge;
-
-	ge.group = &g__;
-	UK_SLIST_INIT(&groups);
-	UK_SLIST_INSERT_HEAD(&groups, &ge, entries);
-}
-
-void setgrent(void)
-{
-	groups_iter = UK_SLIST_FIRST(&groups);
-}
-
-void endgrent(void)
-{
-	setgrent();
-}
-
-struct group *getgrent(void)
-{
-	struct group *res;
-
-	if (groups_iter) {
-		res = groups_iter->group;
-		groups_iter = UK_SLIST_NEXT(groups_iter, entries);
-	} else
-		res = NULL;
-
-	return res;
-
-}
